[dev] refactor the calling order of modules (developing).
