+++ config
@@ -47,7 +47,6 @@ RTMP_DEPS="                                                     \
                 $ngx_addon_dir/ngx_rtmp_record_module.h         \
                 $ngx_addon_dir/ngx_rtmp_gop_cache_module.h      \
                 $ngx_addon_dir/ngx_rtmp_relay_module.h          \
-                $ngx_addon_dir/ngx_rtmp_exec_module.h           \
                 $ngx_addon_dir/ngx_rtmp_streams.h               \
                 $ngx_addon_dir/ngx_rtmp_bitop.h                 \
                 $ngx_addon_dir/ngx_rtmp_proxy_protocol.h        \
